+++ src/Ninja/G1CP/Content/Fixes/Session/fix157_SpeedPotion2Value.d
@@ -28,5 +28,5 @@ func int G1CP_157_SpeedPotion2Value() {
 
     // Replace any assignments to value
     var int symbId; symbId = MEM_GetSymbolIndex("ItFo_Potion_Haste_02");
-    return (G1CP_ReplaceAssignInt(symbId, "C_ITEM.value", 0, Value_Haste1, Value_Haste2) > 0);
+    return (G1CP_ReplaceAssignInt(symbId, 0, "C_ITEM.value", 0, Value_Haste1, Value_Haste2) > 0);
 };